Help text fix: new clarity about the "u" in `umask`..
It stood for "use"! All this time we thought it stood for "user".

First manual to mention umask, for UNIX v7, does not mention
"user", but it does say "Umask sets a mask used whenever a
file is created by [...]" - note "used".

`umask` for "use mask" would be extremely consistent with UNIX
system call naming. "stime" for "set time", "smdate" for "set
modified date, and so on. This can be seen in UNIX v1 through
UNIX v7 manual pages, consistently.

Why "use mask" and not "set mask"? Presumably because "use" is
the more central concept in "set the mask that is used for all
file creation operations that this process does".

Works did not start to say or imply that the "u" stands for
"user" until 1985, according to this seemingly trust-worthy
answer: https://unix.stackexchange.com/a/583484/97172

No one has a good explanation for why or how the "u" in
"umask" would stand for "user".

"User" never quite made sense anyway: how is it a "user" mask
when it is scoped to a process?

That's why I used "umask" as a cryptic noun in the help text.
I could not see a clear way to make "user file creation mode
mask" make sense, but lacked any clarity on how the "u" made
sense if I omitted "user".

Now that it's clear, I am happier with this new wording and
parameter naming in the help text.