Fix misguided usage of "umask" as a noun
Now that we know the "u" stands for "use" (covered in
earlier commit), it is obvious that "umask" is not a
noun, and the actual noun to use everywhere is "mask".

Arguably this is a breaking change, except I hadn't
yet formally committed to the error strings as part
of the API, and I do not know of any users of this
tool besides myself so far.